+++ src/domain/email-prototypes/plainjs/template/index.js
