+++ src/domain/email-prototypes/plainjs/typography/config.js
