+++ src/domain/email-prototypes/plainjs/typography/literals/image.js
