+++ src/domain/email-prototypes/plainjs/typography/literals/italic.js
