+++ src/domain/email-prototypes/plainjs/typography/literals/mainTitle.js
