+++ src/domain/email-prototypes/plainjs/typography/literals/separator.js
