+++ src/domain/email-prototypes/plainjs/typography/literals/strong.js
