+++ src/domain/replacer-class/configuration-nmtg.js
@@ -1,1 +1,13 @@
 //--------------------
+// import {
+//     replaceMarkdown,
+    
+//   } from '../domain/replace-markdown/replace-md';
+
+//  import Replacer from '../../parserMDHtml/replace';
+
+function configureReplacer(state) {
+
+}
+
+export default configureReplacer;
