+++ src/domain/replacer-class/configuration-react.js
@@ -1,1 +1,15 @@
 //--------------------
+
+//--------------------
+// import {
+//     replaceMarkdown,
+    
+//   } from '../domain/replace-markdown/replace-md';
+
+//  import Replacer from '../../parserMDHtml/replace';
+
+function configureReplacer(state) {
+    
+}
+
+export default configureReplacer;
